Separate metadata and difficulty settings
